Move datatype to `TxData` to solve circular module dependency
